Trust pilot widget implementation
