+++ .gitignore
@@ -53,7 +53,7 @@ typings/
 
 # dotenv environment variables file
 .env
-.env*
+.env.development 
 
 # gatsby files
 .cache/
