--- src/images/svg/trustpilot/stars/index.js
@@ -1,11 +1,0 @@
-import Star1 from './star-1.svg'
-import Star1Half from './star-1.5.svg'
-import Star2 from './star-2.svg'
-import Star2Half from './star-2.5.svg'
-import Star3 from './star-3.svg'
-import Star3Half from './star-3.5.svg'
-import Star4 from './star-4.svg'
-import Star4Half from './star-4.5.svg'
-import Star5 from './star-5.svg'
-
-export default { Star1, Star1Half, Star2, Star2Half, Star3, Star3Half, Star4, Star4Half, Star5 }
